project/api/requirements.txt
certifi==2018.8.24
chardet==3.0.4
Django==2.2.1
djangorestframework==3.9.4
httplib2==0.12.3
idna==2.7
Pillow==5.3.0
pipenv==2018.7.1
psycopg2==2.7.5
psycopg2-binary==2.7.6.1
python-congress==0.3.5
pytz==2018.5
requests==2.20.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.23
virtualenv==16.5.0
virtualenv-clone==0.5.3
